requirements.txt
absl-py==0.13.0
astunparse==1.6.3
atari-py==0.2.9
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
click==8.1.2
cloudpickle==1.6.0
cycler==0.10.0
dnspython==2.2.1
email-validator==1.1.3
et-xmlfile==1.1.0
finplot==1.7
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.34.1
gym==0.19.0
h5py==3.1.0
idna==2.10
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.4.2
mpl-finance==0.10.1
multitasking==0.0.9
mysql-client==0.0.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.58
openpyxl==3.0.7
opt-einsum==3.3.0
pandas==1.3.0
pandas-datareader==0.10.0
Pillow==8.3.1
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
SQLAlchemy==1.4.35
stable-baselines3==1.3.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.2.0
torch==1.10.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==3.0.1
WTFormsValidation==0.1.0
XlsxWriter==1.4.4
yfinance==0.1.63
zipp==3.8.0
